+++ CHANGELOG.md
@@ -1,11 +1,14 @@
  == CHANGELOG ==
 
+#### NifSkope-2.0.dev9-20240422
+
 * Implemented the Update Bounds and Update All Bounds mesh spells for Starfield. Casting these also updates the vertex and triangle counts on meshes.
 * Updating the bounds of Fallout 76 shapes now also recalculates the bounding box.
 * Fixed duplicate "Update Bounds" entries in the spells menu with models from Skyrim or newer games.
 * Improved bounding sphere calculation, using code from [Miniball](https://github.com/hbf/miniball) by Martin Kutz, Kaspar Fischer and Bernd Gärtner.
 * Implemented rendering Starfield and Fallout 76 bounding boxes and spheres.
 * Added support for reading version 21 Fallout 76 material files. The new data fields (two texture paths and an unknown byte in effect materials) are currently ignored.
+* Fixed emissive multiple controller on Fallout 76 materials.
 * The "Max Filepath" header field is now read as an array of unknown bytes from Starfield NIF files, to avoid converting binary data to UTF-8 on saving the model.
 
 #### NifSkope-2.0.dev9-20240412